refactor: Create Inline Comments component
